requirements.txt
fill_voids==2.0.6
gradio==4.44.0
numpy==2.1.1
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pydicom==3.0.1
scikit_image==0.19.3
scipy==1.14.1

SimpleITK==2.4.0

tqdm==4.66.2
zy_common_tools==0.1.3
